requirements.txt
# Core dependencies
httpx>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Scheduling
apscheduler>=3.10.4

# HTML parsing
beautifulsoup4>=4.12.3
html5lib>=1.1

# AI - Groq
groq>=0.4.2

# Timezone handling
pytz>=2024.1

# Logging
colorlog>=6.8.2
